Fixed Restaurant Detail Screen Food Lists and Navigations

+++ app/src/main/java/com/pilekumatlar/foodorderingapp/fragments/addingfragments/FragmentAddFood.kt
@@ -40,7 +40,7 @@ class FragmentAddFood : Fragment(R.layout.fragment_add_food_item) {
             var foodDescription: String = binding.editTextFoodDescription.text.toString()
             var foodPrice: String = binding.editTextFoodPrice.text.toString()
 
-            val foodInformations = FoodItem(foodName, foodDescription, foodPrice.toInt())
+            val foodInformations = FoodItem(foodName, foodDescription, foodPrice)
 
             mFirestore.collection("Foods")
                 .document()
